@@ -3,15 +3,18 @@
 class MultiSessionCest
 {
 
+
     /**
-     * Checking that different sessions are taken
+     * @param \WebGuy $I
+     * @param $scenario
+     * @return void
      */
-    public function dontSeeVisualChangesInMultiSessions(WebGuy $I, $scenario)
+    public function dontSeeVisualChangesInMultiSessions(WebGuy $I, $scenario): void
     {
         $I->amOnPage("/multiSession.html");
         $I->dontSeeVisualChanges("block", ".block");
 
-        $friend = $I->haveFriend('friend');
+        $friend = $I->haveFriend('friend', 'WebGuy');
         $friend->does(function (WebGuy $I) {
             $I->amOnPage("/multiSessionChanged.html");
             $I->dontSeeVisualChanges("blockInAnotherSession", ".block");
@@ -20,4 +23,5 @@ class MultiSessionCest
         $I->amOnPage("/multiSession.html");
         $I->dontSeeVisualChanges("block", ".block");
     }
+
 }
